crawlerhistory2/requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
selenium==4.7.2
PyYAML==6.0